backend/streetdrop-notification/src/main/java/com/depromeet/security/provider/ApiKeyAuthProvider.java
package com.depromeet.security.provider;

import com.depromeet.error.code.GlobalErrorCode;
import com.depromeet.error.exceptions.UnauthorizedException;
import org.springframework.beans.factory.annotation.Value;
import org.springframework.security.authentication.AuthenticationProvider;
import org.springframework.security.core.Authentication;
import org.springframework.security.core.AuthenticationException;
import org.springframework.security.web.authentication.preauth.PreAuthenticatedAuthenticationToken;

public class ApiKeyAuthProvider implements AuthenticationProvider {

    @Value("${auth.secret-key}")
    private String secretKey;

    @Override
    public Authentication authenticate(Authentication authentication) throws AuthenticationException {
        String principal = (String) authentication.getPrincipal();
        if (!secretKey.equals(principal))
            throw new UnauthorizedException(GlobalErrorCode.UNAUTHORIZED);
        authentication.setAuthenticated(true);
        return authentication;
    }

    @Override
    public boolean supports(Class<?> authentication) {
        return PreAuthenticatedAuthenticationToken.class.isAssignableFrom(authentication);
    }

}
